backend/requirements.txt
astroid==2.5.1
certifi==2020.12.5
chardet==4.0.0
Click==7.0
colorama==0.4.4
ecdsa==0.16.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
future==0.18.2
idna==2.10
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycryptodome==3.3.1
pylint==2.7.2
python-dotenv==0.15.0
python-jose-cryptodome==1.3.2
requests==2.25.1
six==1.12.0
SQLAlchemy==1.3.3
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.11.1
